The aplication for phase 1 is finished

+++ server/routes/socket.js
@@ -7,8 +7,8 @@ module.exports = {
 
     connect: function(io,PORT){
       const rooms = [
-        { id: 1, name: 'Room 1', channels: [], users: [] },
-        { id: 2, name: 'Room 2', channels: [], users: [] },
+        { id: 1, name: 'Room 1', channels: [{id: 1, name: 'Channel 2'},{id: 2, name: 'Channel 1'}], users: [] },
+        { id: 2, name: 'Room 2', channels: [{id: 1, name: 'Channel 3'},{id: 2, name: 'Channel 4'}], users: [] },
         { id: 3, name: 'Room 3', channels: [], users: [] },
       ];
   
@@ -19,6 +19,11 @@ module.exports = {
                 console.log("this is the testing",JSON.stringify(rooms))
               }
             socket.on('createRoom', (roomName) => {
+              const existingRoom = rooms.find((room) => room.name === roomName);
+              if (existingRoom) {
+              
+                socket.emit('roomCreationError', 'Room with the same name already exists.');
+              } else {
               const newRoom = {
                 id: rooms.length +1,
                 name: roomName,
@@ -37,24 +42,27 @@ module.exports = {
             
                 console.log(`Room "${newRoom}" created.`);
                 console.log(rooms);
-              });
+              }});
               socket.on('createChannel', (data) => {
                 const { channelName, roomId } = data;
-        
-               
-                const newChannel = {
-                  name: channelName,
-                
-                };
-        
               
                 const room = rooms.find((room) => room.id === roomId);
                 if (room) {
+                  const newChannel = {
+                    id: room.channels.length + 1, 
+                    name: channelName,
+                  };
+              
                   room.channels.push(newChannel);
+              
                   io.to(roomId).emit('channelCreated', newChannel);
                   sendUpdatedRoomList();
+                } else {
+                  
+                  socket.emit('channelCreationError', 'Room does not exist.');
                 }
               });
+              
             socket.on('joinRoom', (roomId) => {
               const room = rooms.find((room) => room.id === roomId);
               if (room) {
@@ -76,19 +84,71 @@ module.exports = {
               }
             });
             
-              socket.on('addUser', (newUser) => {
               
-                userData.people.push(newUser);
+              socket.on('addUser', (newUser) => {
                 
+                const existingUserByName = userData.people.find((user) => user.username === newUser.username);
+              
                 
-                fs.writeFileSync(userDataFilePath, JSON.stringify(userData), 'utf8');
-
-                
-                io.emit('userAdded', newUser);
-
-                console.log(`User "${newUser.username}" added.`);
-                console.log(userData);
+                const existingUserByEmail = userData.people.find((user) => user.email === newUser.email);
+              
+                if (existingUserByName) {
+                  
+                  socket.emit('userCreationError', 'Username already exists.');
+                } else if (existingUserByEmail) {
+                  
+                  socket.emit('userCreationError', 'Email already exists.');
+                } else {
+              
+                  newUser.id = userData.people.length + 1
+                  userData.people.push(newUser);
+              
+                  fs.writeFileSync(userDataFilePath, JSON.stringify(userData), 'utf8');
+              
+                  io.emit('userAdded', newUser);
+              
+                  console.log(`User "${newUser.username}" added.`);
+                  console.log(userData);
+                }
               });
+              
+              //Add User To Room
+              socket.on('addUserToRoom', ({ userId, roomId }) => {
+                const room = rooms.find((room) => room.id === roomId);
+                if (room) {
+                  if (!room.users.includes(userId)) {
+                    room.users.push(userId);
+                    io.to(roomId).emit('userAddedToRoom', { userId, roomId });
+                  } else {
+                    // Gửi thông báo lỗi nếu người dùng đã có trong phòng
+                    socket.emit('userAddToRoomError', 'User is already in this room.');
+                  }
+                } else {
+                  // Gửi thông báo lỗi nếu phòng không tồn tại
+                  socket.emit('userAddToRoomError', 'Room does not exist.');
+                }
+              });
+              // Trong file socket.js
+              socket.on('addChannelToRoom', (data) => {
+                const { channelName, roomId } = data;
+
+                const room = rooms.find((room) => room.id === roomId);
+                if (room) {
+                  const newChannel = {
+                    name: channelName,
+                  };
+
+                  room.channels.push(newChannel);
+
+                  io.to(roomId).emit('channelCreated', newChannel);
+                  sendUpdatedRoomList();
+                } else {
+                  // Gửi thông báo lỗi nếu phòng không tồn tại
+                  socket.emit('channelCreationError', 'Room does not exist.');
+                }
+              });
+
+              
 
      
               socket.on('leaveRoom', (roomId) => {
@@ -108,33 +168,67 @@ module.exports = {
               });
               socket.on('deleteUser', (userId) => {
                 console.log(userData)
-                // Tìm người dùng cần xóa trong danh sách người dùng
+             
                 const userIndex = userData.people.findIndex((user) => user.id === userId);
               
                 if (userIndex !== -1) {
-                  // Xóa người dùng khỏi danh sách
+                  
                   userData.people.splice(userIndex, 1);
               
-                  // Ghi lại dữ liệu người dùng vào tệp
+                 
                   fs.writeFileSync(userDataFilePath, JSON.stringify(userData), 'utf8');
               
-                  // Gửi thông báo xóa người dùng thành công cho máy khách
+                
                   socket.emit('userDeleted', userId);
               
                   console.log(`User with ID ${userId} has been deleted.`);
                   console.log(userData)
                 } else {
-                  // Gửi thông báo lỗi nếu không tìm thấy người dùng
+                  
                   socket.emit('userDeleteError', 'User not found.');
                 }
               });
               socket.on('getUsersList', () => {
-                // Lấy danh sách người dùng từ dữ liệu của bạn (userData)
+              
                 const userList = userData.people;
             
-                // Gửi danh sách người dùng cho client
+               
                 socket.emit('usersList', userList);
               });
+
+              socket.on('addUserToChannel', ({ channelName, roomId, userId }) => {
+                // Tìm phòng theo roomId
+                const room = rooms.find((room) => room.id === roomId);
+                
+                if (room) {
+                  
+                  const channel = room.channels.find((ch) => ch.name === channelName);
+              
+                  if (channel) {
+                    
+                    if (!channel.users.includes(userId)) {
+                     
+                      channel.users.push(userId);
+              
+                 
+                      io.to(channelName).emit('userAddedToChannel', { channelName, userId });
+              
+                    
+                      sendUpdatedRoomList();
+                    } else {
+                      
+                      socket.emit('userAddToChannelError', 'User is already in this channel.');
+                    }
+                  } else {
+                
+                    socket.emit('channelAddToChannelError', 'Channel does not exist in this room.');
+                  }
+                } else {
+                 
+                  socket.emit('roomAddToChannelError', 'Room does not exist.');
+                }
+              });
+
